refactor: split waveform decode readers
